requirements.txt
aiodns==3.0.0
aiohttp==3.8.1
aiosignal==1.2.0
appdirs==1.4.4
appnope==0.1.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
async-timeout==4.0.2
attrs==22.1.0
autobahn==21.2.1
Automat==22.10.0
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
binance-futures-connector==1.10.0
bleach==4.1.0
ccxt==1.77.81
certifi==2021.10.8
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.0.12
constantly==15.1.0
cryptography==38.0.4
dateparser==1.1.1
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
executing==0.8.3
frozenlist==1.3.0
hyperlink==21.0.0
idna==2.10
importlib-resources==5.6.0
incremental==22.10.0
ipykernel==6.11.0
ipython==8.2.0
ipython-genutils==0.2.0
jedi==0.18.1
Jinja2==3.1.1
jsonschema==4.4.0
jupyter-client==7.2.1
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mistune==0.8.4
multidict==6.0.2
nbclient==0.5.13
nbconvert==6.4.5
nbformat==5.2.0
nest-asyncio==1.5.4
notebook==6.4.10
numpy==1.22.3
packaging==21.3
pandas==1.4.1
pandas-ta==0.3.14b0
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
playsound==1.3.0
prometheus-client==0.13.1
prompt-toolkit==3.0.28
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycares==4.1.2
pycparser==2.21
Pygments==2.11.2
pyOpenSSL==19.0.0
pyparsing==3.0.7
pyrsistent==0.18.1
python-binance==1.0.15
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pyzmq==22.3.0
regex==2022.3.2
requests==2.25.1
Send2Trash==1.8.0
service-identity==21.1.0
six==1.16.0
soupsieve==2.3.1
stack-data==0.2.0
terminado==0.13.3
testpath==0.6.0
tornado==6.1
traitlets==5.1.1
Twisted==22.2.0
txaio==22.2.1
typing_extensions==4.4.0
tzdata==2022.1
tzlocal==4.1
ujson==5.1.0
urllib3==1.26.9
wcwidth==0.2.5
webencodings==0.5.1
websockets==9.1
yarl==1.7.2
zipp==3.7.0
zope.interface==5.5.2
